+++ R/QueryNamespace.R
@@ -79,7 +79,6 @@ QueryNamespace <- R6::R6Class(
       if (!is.null(tableSpecification)) {
         self$addTableSpecification(tableSpecification)
       }
-
       replaceVars <- list(...)
 
       if (length(replaceVars)) {
@@ -87,6 +86,8 @@ QueryNamespace <- R6::R6Class(
           self$addReplacementVariable(k, replaceVars[[k]])
         }
       }
+
+      self
     },
 
     #' Set Connection Handler
@@ -95,10 +96,11 @@ QueryNamespace <- R6::R6Class(
     setConnectionHandler = function(connectionHandler) {
       checkmate::assertR6(connectionHandler, "ConnectionHandler")
       private$connectionHandler <- connectionHandler
+      invisible(NULL)
     },
 
-    #' Get connection handler
-    #' @description get connection handler obeject or throw error if not set
+        #' Get connection handler
+        #' @description get connection handler obeject or throw error if not set
     getConnectionHandler = function() {
       if (is.null(private$connectionHandler)) {
         stop("ConnectionHandler not set")
@@ -120,6 +122,7 @@ QueryNamespace <- R6::R6Class(
       }
 
       private$replacementVars$set(key, value)
+      invisible(NULL)
     },
 
     #' add table specification
@@ -141,6 +144,7 @@ QueryNamespace <- R6::R6Class(
 
         self$addReplacementVariable(tableName, replacementVar, replace = replace)
       }
+      invisible(NULL)
     },
 
     #' Render
@@ -184,11 +188,72 @@ QueryNamespace <- R6::R6Class(
       ch$executeSql(sql)
     },
 
-    #' get vars
-    #' @description
-    #' returns full list of variables that will be replaced
+        #' get vars
+        #' @description
+        #' returns full list of variables that will be replaced
     getVars = function() {
       return(private$replacementVars$as_list())
     }
   )
 )
+
+#' Create query namespace
+#'
+#' @export
+#' @description
+#' Create a QueryNamespace instance from either a connection handler or a connectionDetails object
+#' Allows construction with various options not handled by QueryNamespace$new
+#'
+#' Note - currently not supported is having multiple table prefixes for multiple table namespaces
+#'
+#' @inheritParams QueryNamespace
+#' @inheritParams ConnectionHandler
+#' @param resultModelSpecificationPath (optional) csv file or files for tableSpecifications - must conform to table spec
+#'                                     format.
+createQueryNamespace <- function(connectionDetails = NULL,
+                                 connectionHandler = NULL,
+                                 usePooledConnection = FALSE,
+                                 tableSpecification = NULL,
+                                 resultModelSpecificationPath = NULL,
+                                 tablePrefix = "",
+                                 snakeCaseToCamelCase = TRUE,
+                                 ...) {
+
+  checkmate::assertClass(connectionDetails, "ConnectionDetails", null.ok = TRUE)
+  checkmate::assertClass(connectionHandler, "ConnectionHandler", null.ok = TRUE)
+  checkmate::assertLogical(usePooledConnection)
+  checkmate::assertDataFrame(tableSpecification, null.ok = TRUE)
+  checkmate::assertString(tablePrefix)
+
+  if (!is.null(tableSpecification))
+    assertSpecificationColumns(colnames(tableSpecification))
+
+  if (is.null(connectionDetails) && is.null(connectionHandler))
+    stop("Must provide ConnectionDetails or ConnectionHandler instance")
+
+  if (!is.null(resultModelSpecificationPath)) {
+    if (is.null(tableSpecification))
+      tableSpecification <- data.frame()
+
+    # Create a merged specification from all spec files provided.
+    tableSpecification <- lapply(resultModelSpecificationPath,
+                                 loadResultsDataModelSpecifications) %>%
+      dplyr::bind_rows(tableSpecification)
+
+  }
+
+  if (!is.null(connectionDetails)) {
+    if (usePooledConnection)
+      connectionHandler <- PooledConnectionHandler$new(connectionDetails)
+    else
+      connectionHandler <- ConnectionHandler$new(connectionDetails)
+  }
+  connectionHandler$snakeCaseToCamelCase <- snakeCaseToCamelCase
+
+  qns <- QueryNamespace$new(connectionHandler,
+                            tableSpecification = tableSpecification,
+                            tablePrefix = tablePrefix,
+                            ...)
+
+  return(qns)
+}